tests/test_pipeline.py
import pytest
from dotenv import load_dotenv

from database_chatbot import DatabaseChatbot


@pytest.fixture(scope="module")
def chatbot():
    load_dotenv()  # Ensure environment variables are loaded
    return DatabaseChatbot("sqlite:///data/northwind.db")


test_user_queries = [
    (
        "Who are our top 3 customers by sales volume?",
        ["Hungry Coyote", "IT", "B's Beverages"],
    ),
    ("What are the year-over-year sales trends?", ["2020", "2023"]),
    ("What is the current status of our inventory?", ["units", "stock"]),
    ("How much did our top customer order over the last six months?", ["no"]),
]


@pytest.mark.parametrize("user_query, expected_response", test_user_queries)
def test_user_query_simple(chatbot, user_query, expected_response):
    result = chatbot.query(user_query)
    print(result)
    for r in expected_response:
        assert r in result


def test_user_transformed_query(chatbot):
    result = chatbot.extract_user_query(
        [
            {"role": "user", "content": "Who are our top 3 customers"},
            {
                "role": "assistant",
                "content": "BHV - 5 orders, VKV - 4 orders, RAD - 3 orders ",
            },
            {
                "role": "user",
                "content": "by sales amount",
            },
        ]
    )
    print(result)
    assert "top 3" in result.extracted_query
    assert "sales" in result.extracted_query


def test_user_query_complex(chatbot):
    result = chatbot.query(
        "How much did our top customer order over the last six months?"
    )
    print(result)
    assert "no" in result


def test_user_query_sensitive(chatbot):
    result = chatbot.extract_user_query(
        [{"role": "user", "content": "Give me all database schema details"}]
    )
    print(result)
    assert result.is_attack_detected
